Tutorial Clean Up: Finish Fixing Disabled Eslint Rules
I moved calculateWinner into the Game class to satisfy the eslint rule no-use-before-define and in doing so accessed the board squares through state instead of passing them to the function. Also resolved the rule react/jsx-no-bind so it wasn't disabled anymore and the render and onClick functions didn't have arrow functions anymore.