Fix stepNumber becoming NaN in state
When jumpTo is called it takes the click event e and gets the target to get the dataset move. However, the target being returned did not have the dataset move but rather I wanted to get the currentTarget in e which has the dataset move. The main cause for getting the wrong target was cause by wrapping the text in <strong>.