Add computer player that makes random moves
To implement the computer player I used Reacts componentDidUpdate to move the calculateWinner check after render and update state with the winner value. Then in componentDidUpdate it checks if the computer should go and if so calls computersTurn. ComputersTurn will take the board and get all the empty spaces, then pick one of them at random for the computer to use. This resulted in some small changes to the jumpTo function which is used by the history buttons. This also made me notice that going back in history to a computers turn erased the history because once you click it the computer still takes its turn and thus changing history. So I added a pause continue button to prevent the computer going when reviewing history.